+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - The DNA indices where spaced beads have been inserted are now saved to `post_processing_parameters.py`.
+- New `extension` parameter controls the relative extension of the DNA (for `Line` configuration only).
 
 ### Fixed